:seedling:注册页面

+++ src/main/java/com/summer/weibo/Repository/UserRepository.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
  * @Version 1.0
  */
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 }
